+++ jax/_src/pallas/mosaic/core.py
@@ -22,6 +22,7 @@ import functools
 from typing import Any, ClassVar, Literal
 
 import jax
+from jax._src import config
 from jax._src import core as jax_core
 from jax._src import dtypes
 from jax._src import util
@@ -44,6 +45,17 @@ AbstractMemoryRef = pallas_core.AbstractMemoryRef
 no_block_spec = pallas_core.no_block_spec
 _convert_block_spec_to_block_mapping = pallas_core._convert_block_spec_to_block_mapping
 split_list = util.split_list
+
+_ENABLE_RUNTIME_ASSERT = config.bool_state(
+    "jax_pallas_enable_runtime_assert",
+    default=False,
+    help=(
+        "If set, enables runtime assertions in the kernel via checkify.check."
+        " Otherwise, runtime asserts will be ignored unless functionalized"
+        " using checkify.checkify."
+    ),
+)
+
 
 @dataclasses.dataclass(frozen=True)
 class TPUCompilerParams(pallas_core.CompilerParams):
@@ -203,3 +215,7 @@ def create_tensorcore_mesh(axis_name: str) -> pallas_core.PallasMesh:
       np.array([TensorCore(i) for i in range(num_cores)]),
       [axis_name],
   )
+
+def runtime_assert_enabled() -> bool:
+  """Returns whether runtime asserts are enabled."""
+  return _ENABLE_RUNTIME_ASSERT.value